v1.6! so many changes:
Saved vehicles works well. Updates are instant in both the popup and the
content script. It would be cool if the content script would auto-reload
the page to reflect updated values. An issue for the future.

Some styling changes. Hope to have the vehicle interface working better
soon. Finally FINALLY wrangled the storage API.